netconf stuff moved
