+++ main.py
@@ -4,7 +4,6 @@ Simple (and portable) NETCONF client
 Copyright (c) 2024 Ejub Šabić <ejub1946@outlook.com>
 """
 import os
-import json
 import re
 import sys
 import argparse
@@ -19,20 +18,18 @@ import signal
 import logging
 import threading
 import tkinter as tk
-from tkinter import Menu, END, filedialog, messagebox, Listbox, \
-    Toplevel, TclError
+from tkinter import Menu, END, filedialog, messagebox, Toplevel, TclError
 from xml.dom.minidom import parseString
 import psutil
 from zeroconf import ServiceBrowser, Zeroconf
 from PIL import Image, ImageTk
 import customtkinter as ctk
-from ncclient import manager
-from ncclient.transport.errors import AuthenticationError, SSHError
 from ncclient.xml_ import to_ele
 from lxml.etree import XMLSyntaxError, DocumentInvalid
 from tkinterweb import HtmlFrame
 
-from simple_netconf_client.gui.Dialogs import LicenseDialog, UsageDialog, AboutDialog
+from simple_netconf_client.gui.Dialogs import LicenseDialog, UsageDialog, AboutDialog, ScanResultsDialog
+from simple_netconf_client.network.Netconf import ConfigManager, ZeroconfListener, NetconfConnection
 
 APP_TITLE = "Simple NETCONF Client"
 WELCOME = "Welcome to the NETCONF client!\n\n" \
@@ -51,8 +48,6 @@ WELCOME = "Welcome to the NETCONF client!\n\n" \
     " - Reboot\n" \
     " - Verify device comes back up\n" \
     " - Verify upgrade \"took\", check version with Get Status\n"
-ctk.set_appearance_mode("System")
-ctk.set_default_color_theme("blue")
 
 RPC_SYSTEM_RESTART = """<system-restart xmlns="urn:ietf:params:xml:ns:yang:ietf-system"/>"""
 RPC_FACTORY_RESET = """<factory-reset xmlns="urn:ietf:params:xml:ns:yang:ietf-factory-default"/>"""
@@ -69,6 +64,10 @@ RPC_GET_OPER = """<get-data xmlns="urn:ietf:params:xml:ns:yang:ietf-netconf-nmda
     </subtree-filter>
 </get-data>
 """
+
+ctk.set_appearance_mode("System")
+ctk.set_default_color_theme("blue")
+
 def resource_path(relative_path):
     try:
         base_path = sys._MEIPASS
@@ -76,151 +75,6 @@ def resource_path(relative_path):
         base_path = os.path.abspath(".")
 
     return os.path.join(base_path, relative_path)
-
-class ConfigManager:
-    def __init__(self, filename='.netconf_config.json'):
-        self.filename = filename
-        self.filepath = self._get_file()
-        self.default_cfg = {
-            'addr': '',
-            'port': 830,
-            'user': "admin",
-            'pass': '',
-            'ssh-agent': True,
-            'theme': "System",
-            'zoom': "100%",
-            'server_iface': 'virbr0',
-            'server_enabled': False,
-            'server_path': '',
-            'server_port': 8080
-        }
-        self.cfg = self.default_cfg.copy()
-        self.load()
-
-    def _get_file(self):
-        home_dir = os.path.expanduser('~')
-        return os.path.join(home_dir, self.filename)
-
-    def save(self):
-        with open(self.filepath, 'w') as file:
-            json.dump(self.cfg, file)
-
-    def load(self):
-        if os.path.exists(self.filepath):
-            with open(self.filepath, 'r') as file:
-                self.cfg = json.load(file)
-        # Merge default config with loaded config
-        self._merge_defaults()
-
-    def _merge_defaults(self):
-        for key, value in self.default_cfg.items():
-            if key not in self.cfg:
-                self.cfg[key] = value
-
-
-class NetconfConnection:
-    def __init__(self, cfg, app):
-        self.cfg = cfg
-        self.app = app
-        self.manager = None
-
-    def __enter__(self):
-        try:
-            host = self.cfg['addr']
-            port = self.cfg['port']
-            self.app.status(f"Connecting to {host} port {port}, please wait ...")
-            self.manager = manager.connect(host=host,
-                                           port=port,
-                                           username=self.cfg['user'],
-                                           password=self.cfg['pass'],
-                                           hostkey_verify=False,
-                                           allow_agent=self.cfg['ssh-agent'],
-                                           timeout=30)
-            logging.info("Connected to NETCONF server %s.", host)
-            return self.manager
-        except AuthenticationError as err:
-            self.app.error(f"Authentication with {host} failed: {err}")
-        except SSHError as err:
-            self.app.error(f"SSH connection to {host}:{port} failed: {err}")
-        except Exception as err:
-            self.app.error(f"An unexpected error occurred: {err}")
-            raise err
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if self.manager is not None:
-            self.manager.close_session()
-            logging.info("Disconnected from NETCONF server")
-
-
-class ZeroconfListener:
-    def __init__(self, app):
-        self.app = app
-        self.devices = {}
-
-    def add_service(self, zeroconf, type, name):
-        info = zeroconf.get_service_info(type, name)
-        if info:
-            addresses = [socket.inet_ntoa(addr) for addr in info.addresses]
-            hostname = info.server if info.server else name
-            for address in addresses:
-                self.devices[name] = (hostname, address, info.port)
-
-            self.app.update_device_list(self.devices)
-
-    def update_service(self, zeroconf, type, name):
-        pass
-
-    def remove_service(self, zeroconf, type, name):
-        if name in self.devices:
-            del self.devices[name]
-            self.app.update_device_list(self.devices.values())
-
-
-class ScanResultsDialog(ctk.CTkToplevel):
-    def __init__(self, parent, devices):
-        super().__init__(parent)
-        self.title("mDNS-SD - Select NETCONF Device")
-        self.geometry("400x300")
-        self.devices = devices
-        self.selected_device = None
-
-        self.listbox = Listbox(self)
-        self.listbox.pack(fill="both", expand=True, padx=10, pady=10)
-        self.update_listbox()
-
-        self.cancel_button = ctk.CTkButton(self, text="Cancel",
-                                           command=self.on_cancel)
-        self.cancel_button.pack(side="left", padx=10, pady=10)
-
-        self.ok_button = ctk.CTkButton(self, text="OK",
-                                       command=self.on_ok)
-        self.ok_button.pack(side="right", padx=10, pady=10)
-
-        # Allow closing with common key bindings
-        self.bind("<Control-w>", lambda event: self.destroy())
-        self.bind("<Escape>", lambda event: self.destroy())
-        # Bring frame into foreground and focus it when it becomes visible
-        self.bind("<Map>", self.on_map)
-
-    def on_map(self, event):
-        self.lift()
-        self.focus_force()
-
-    def update_listbox(self):
-        self.listbox.delete(0, END)
-        for name, ip, port in self.devices:
-            name = name.rstrip('.')
-            self.listbox.insert('end', f"{name} - {ip}:{port}")
-
-    def on_ok(self):
-        selection = self.listbox.curselection()
-        if selection:
-            self.selected_device = self.devices[selection[0]]
-        self.destroy()
-
-    def on_cancel(self):
-        self.destroy()
-
 
 class App(ctk.CTk):
     def __init__(self):
